Add aoi_data_directory fixture to additional tests where the ntv2.0.gsb file can cause tests to fail
